requirements.txt
adafruit-circuitpython-rgb-display
numpy
Pillow
requests
spidev
systemd